Refactor: UserDetails 구현체 변경
